Undid SSO stuff. Added ApiController for public use. Updated Question.kt
It seemed to fit better using enumerated types rather than numerical types in Question and QuestionSet. Additionally, realized Question needed a type parameter so clients know how to present the answers and inputs. The correctAnswerId field is meant to show which of the answers is correct. If all answers are correct or any user given answer is correct, this field can be set null.